Revert "refactor: rename polylines to editable polylines"
This reverts commit 3c69289a604277413aadd6b4ea0adca80474ef83.
